Removed User Component.
